Remove set_user_id and some formatting

+++ app/controllers/api/v0/users_controller.rb
@@ -40,11 +40,11 @@ class Api::V0::UsersController < Api::V0::ApplicationController
 
   end
 
-	def suggestions
-		#This should be a "smart" algorithm. For now just returning all users not being followed by current user
-		users = User.where.not(:id => @current_user.following_ids | [@current_user.id])
-		render json: users
-	end
+  def suggestions
+    #This should be a "smart" algorithm. For now just returning all users not being followed by current user
+    users = User.where.not(id: @current_user.following_ids | [@current_user.id])
+    render json: users
+  end
 
   def show
     user = User.find(params[:id])
@@ -66,9 +66,7 @@ class Api::V0::UsersController < Api::V0::ApplicationController
 
   def clear_session
     auth = Auth.find_by(access_token: request.headers["HTTP_ACCESS_TOKEN"])
-    if auth
-      auth.destroy!
-    end
+    auth.destroy! if auth
   end
 
   def friendships
